feat: enhanced Aqueduct ETL with AI scoring, CDM documentation, and supporting assets
Backend:
- Enhanced EtlSuggestionScorer with comprehensive healthcare domain aliases
  and multi-tier semantic matching (exact, suffix, alias, fuzzy)
- EtlSuggestionService now returns AI-generated transformation logic
- EtlFieldMappingController field suggestion endpoint
- CDM schema config enriched with ETL conventions, descriptions, and FK metadata
- API route for field-level suggestions

Frontend:
- CDM schema TypeScript definitions synchronized with backend
- ETL API layer extended with field suggestion types and hooks
- EtlToolsPage passes CDM metadata to mapping components

Infrastructure:
- Darkstar R prediction endpoint additions
- R Dockerfile updates

Documentation & Data:
- IRSF genotype-severity research paper with figures and statistical results
- BlackRabbit synthetic outpatient EHR extract (DDL, inserts, generator)
- CDM documentation scraper script
- PLP connection fix handoff document
- Publish page study-driven redesign spec
- Planning state updates (Phase 12 complete)

+++ backend/app/Services/Analysis/Features/MeasurementFeatureBuilder.php
@@ -27,16 +27,16 @@ class MeasurementFeatureBuilder implements FeatureBuilderInterface
                 COALESCE(c2.concept_name, 'Unknown') AS concept_name,
                 COUNT(DISTINCT m.person_id) AS person_count,
                 ROUND(
-                    100.0 * COUNT(DISTINCT m.person_id) /
-                    NULLIF((SELECT COUNT(DISTINCT subject_id) FROM {@cohortTable} WHERE cohort_definition_id = {$cohortDefinitionId}), 0),
+                    (100.0 * COUNT(DISTINCT m.person_id) /
+                    NULLIF((SELECT COUNT(DISTINCT subject_id) FROM {@cohortTable} WHERE cohort_definition_id = {$cohortDefinitionId}), 0))::numeric,
                     2
                 ) AS percent_value,
-                ROUND(AVG(CASE WHEN m.value_as_number IS NOT NULL THEN m.value_as_number END), 4) AS mean_value,
-                ROUND(STDDEV(CASE WHEN m.value_as_number IS NOT NULL THEN m.value_as_number END), 4) AS stddev_value,
+                ROUND(AVG(CASE WHEN m.value_as_number IS NOT NULL THEN m.value_as_number END)::numeric, 4) AS mean_value,
+                ROUND(STDDEV(CASE WHEN m.value_as_number IS NOT NULL THEN m.value_as_number END)::numeric, 4) AS stddev_value,
                 MIN(CASE WHEN m.value_as_number IS NOT NULL THEN m.value_as_number END) AS min_value,
                 MAX(CASE WHEN m.value_as_number IS NOT NULL THEN m.value_as_number END) AS max_value,
                 ROUND(
-                    PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY CASE WHEN m.value_as_number IS NOT NULL THEN m.value_as_number END),
+                    PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY CASE WHEN m.value_as_number IS NOT NULL THEN m.value_as_number END)::numeric,
                     4
                 ) AS median_value
             FROM {@cohortTable} c
